Move ProgressState into its own file
This didn't really belong in ConsoleAppender.scala anymore.
